update to float64; fix wrong args in rmsd error

+++ src/regression/dataset.py
@@ -16,8 +16,8 @@ class SimulatedDataset(Dataset):
     """
     def __init__(self, path: str, svd: bool):
         with np.load(path) as data:
-            self.cloud = torch.as_tensor(data["cloud"], dtype=torch.float32)
-            self.quat  = torch.as_tensor(data["quat"], dtype=torch.float32)
+            self.cloud = torch.as_tensor(data["cloud"], dtype=torch.float64)
+            self.quat  = torch.as_tensor(data["quat"], dtype=torch.float64)
             self.get_svd = svd
 
     def __len__(self):
@@ -28,7 +28,7 @@ class SimulatedDataset(Dataset):
         curr_quat = self.quat[index]
 
         if self.get_svd:
-            return curr_cloud, torch.as_tensor(direct_SVD(curr_cloud), dtype=torch.float32)
+            return curr_cloud, torch.as_tensor(direct_SVD(curr_cloud), dtype=torch.float64)
         
         return curr_cloud, curr_quat
 
@@ -61,7 +61,7 @@ class ModelNetDataset(Dataset):
 
     def __getitem__(self, index: int):
         random_pick = np.random.randint(len(self.select_files))
-        orig_cloud = torch.as_tensor(F.read_off_file(self.select_files[random_pick]), dtype=torch.float32)
+        orig_cloud = torch.as_tensor(F.read_off_file(self.select_files[random_pick]), dtype=torch.float64)
 
         random_indices = torch.randperm(len(orig_cloud))
         num_points = int(self.num_sample)
@@ -71,18 +71,18 @@ class ModelNetDataset(Dataset):
         curr_rot = generate_random_quat()
         r = R.from_quat(curr_rot)
         rot_mat = r.as_matrix()
-        rot_mat_tensor = torch.as_tensor(rot_mat, dtype=torch.float32)
+        rot_mat_tensor = torch.as_tensor(rot_mat, dtype=torch.float64)
 
         rotate_cloud = torch.matmul(source_cloud, rot_mat_tensor)
         noise = self.sigma*torch.randn_like(source_cloud)
         target_cloud = rotate_cloud + noise
         
-        concatenate_cloud = torch.empty(2, num_points, 3, dtype=torch.float32)
+        concatenate_cloud = torch.empty(2, num_points, 3, dtype=torch.float64)
         
         concatenate_cloud[0,:,:] = source_cloud
         concatenate_cloud[1,:,:] = target_cloud
 
-        return concatenate_cloud, torch.as_tensor(r.as_quat(),dtype=torch.float32)
+        return concatenate_cloud, torch.as_tensor(r.as_quat(),dtype=torch.float64)
 
 class KittiOdometryDataset(Dataset):
     """
